# lib/api-zod/src/generated/types/discardUploadResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export const DiscardUploadResultValue = {
  success: true,
} as const;
export type DiscardUploadResult = typeof DiscardUploadResultValue;